+++ autopilot/NN_Loss.h
@@ -5,6 +5,6 @@
 namespace NN {
 
 [[nodiscard]] auto
-l1Loss(const uint8_t* predicted, const uint8_t* target, const uint16_t size) -> uint32_t;
+l1Loss(const float* predicted, const float* target, const uint16_t size) -> float;
 
 } // namespace NN
